+++ source/frost/storage/perst/IndexSlotsStorage.java
@@ -69,7 +69,7 @@ public class IndexSlotsStorage extends AbstractFrostStorage implements ExitSavab
     @Override
     public boolean initStorage() {
         final String databaseFilePath = buildStoragePath(getStorageFilename()); // path to the database file
-        final int pagePoolSize = getPagePoolSize(SettingsClass.PERST_PAGEPOOLSIZE_INDEXSLOTS);
+        final long pagePoolSize = getPagePoolSize(SettingsClass.PERST_PAGEPOOLSIZE_INDEXSLOTS);
 
         open(databaseFilePath, pagePoolSize, false, true, true);
